requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-mock==3.14.0

#Auth packages
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1